Bug fixes for broken bean compliance. The XML saving routine was not saving/loading things properly. This still needs to be re-tested, but I want to commit so it's not on only one computer.
*NOTE*

For anyone wondering why I chose the bean/XML saving system, it's
actually because I intend to expand it into a better persistence system.
In this early version, everything remains in memory, but that won't last
long. By using the bean persistance systems, this should be a simple
matter to overcome.